'修改删除评论和贴子条件'

+++ app/Http/Controllers/V1/PostCommentController.php
@@ -187,6 +187,10 @@ class PostCommentController extends BaseController
     {
         //
         $postComment = $this->postComment->find($id);
+        if($postComment->user_id!=auth()->id())
+        {
+            abort(422);
+        }
         event(new PostCommentDeleted($postComment));
         $this->postComment->destroy($postComment);
         return $this->response->noContent();
